feat: refactor HelpDrawer and Favorite
